moving a bunch of things around
